Started on the interfaces and implementations (skeleton classes currently) of 2d render targets.

+++ ciri/inc/ciri/gfx/ISamplerState.hpp
@@ -45,7 +45,7 @@ namespace ciri {
 		SamplerWrap::Mode wrapV;
 		SamplerWrap::Mode wrapW;
 		SamplerFilter::Mode filter;
-		float maxAnisotropy;
+		int maxAnisotropy;
 		float borderColor[4];
 		float minLod;
 		float maxLod;
